Extract FlashblocksHarness module
