+++ crates/test-utils/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod accounts;
 pub mod engine;
+pub mod flashblocks_harness;
 pub mod harness;
 pub mod node;
 pub mod tracing;
